refactor(navbar): split follow-link data from its renderer
Exporting followLinks alongside the FollowLinkAnchor component tripped
react-doctor's only-export-components rule, which breaks Fast Refresh and
regressed the quality gate (64 -> 65 errors).
